models handlers using ollama
