--- README.md
@@ -1,30 +1,0 @@
-# afeka_ds_lab_1
-
-This repo is for Afeka data science lab semster 1 2024
-
-## Folder structure
-
-lab\_{_index_}:
-
-- data
-- task.ipynb
-- task.py ( this is the jupyter exported file)
-
-utils:
-
-- pandas operations
-- plot operations
-- etc
-
-## Running and Testing
-
-build a virtual environment, open a terminal and type:\
-`python -m venv venv` \
-Activate your virtual environment:\
-**Linux / Mac**: `source venv/bin activate` \
-**Windows**: `.\venv\bin\activate`
-
-### Help / Contribute
-
-1. to contribute please submit a Pull Request with the feature / bug fix
-2. if you need any assitance you can contact me through me email | Linkedin | Whatsapp
