+++ output/mistral/reasoning1.txt
@@ -1,8 +1,8 @@
 Sentence: After the rain, Sarah grabbed her umbrella before leaving the office.
 Reasoning:
- 1. Observation: It has rained (this can be observed from weather reports or direct observation).
+ 1. The rain occurred - This can be inferred from the fact that Sarah grabs her umbrella, which is typically used during or after rain.
 
-2. Inference: Since it has rained and Sarah is at the office, there's a high likelihood that she would have an umbrella with her since she usually carries one when going out in the rain (based on past behavior).
+2. Sarah was at the office when it rained - This can be deduced because she grabbed her umbrella before leaving the office, indicating that she was there during the rain.
 
-3. Conclusion: Therefore, after the rain, Sarah will likely grab her umbrella before leaving the office because it is necessary to protect herself from further rain. This action is consistent with her previous behavior and the current weather condition.
+3. Sarah left the office after the rain - Since she grabbed her umbrella before leaving, it means she has plans or a need to leave, and since she used her umbrella, it implies that the rain had stopped or was subsiding enough for her to travel.
