+++ output/mistral/reasoning2.txt
@@ -1,8 +1,8 @@
 Sentence: The coffee was too hot to drink, so I added an ice cube.
 Reasoning:
- Reasoning Step 1: The initial observation is that the coffee was too hot to drink, which implies it exceeded a comfortable temperature for consumption.
+ 1. Reasoning Step 1: Identification of Problem - The initial observation is that the coffee was too hot to drink comfortably. This is a problem because one cannot consume liquids at temperatures that may cause discomfort or potentially burns.
 
-Reasoning Step 2: To lower the temperature of the coffee, one possible solution is to add something cold. Ice cubes are commonly used in this situation since they can quickly and effectively cool down liquids.
+2. Reasoning Step 2: Brainstorming Solutions - In search of a solution, you consider several options such as waiting for the coffee to cool down, using a cold spoon, or adding something cold to lower its temperature immediately.
 
-Reasoning Step 3: By adding an ice cube, the temperature of the coffee should decrease as the ice cube melts and mixes with the hot liquid, making it more suitable for drinking. This process would make the coffee easier to consume without causing discomfort or burns.
+3. Reasoning Step 3: Implementation and Evaluation - You decide to add an ice cube to quickly reduce the coffee's temperature without having to wait for it to cool naturally. After adding the ice cube, you evaluate the effectiveness of your solution by testing the temperature of the coffee with your fingertip or using a thermometer (if available). If the coffee is now cool enough to drink comfortably, your problem has been successfully solved.
