+++ output/mistral/reasoning3.txt
@@ -1,13 +1,8 @@
 Sentence: Tom can't go to his sister's wedding because he's studying abroad in Japan.
 Reasoning:
- 1. Tom is currently residing in a place called Japan, as indicated by the phrase "studying abroad in Japan."
+ 1. Identify the Location: The first step is to identify the current location of Tom, which is specified as being "studying abroad in Japan".
 
-2. Traveling from Japan to the location of the sister's wedding requires time and resources, which might be difficult or impossible due to various factors such as flight schedules, cost, visa restrictions, or personal constraints (e.g., study commitments).
+2. Understand Travel Limitations: Given that Tom is studying abroad and not based locally, it follows that he has certain travel limitations due to his studies, visa status, or financial constraints. These limitations make it difficult for him to travel back to where his sister's wedding is taking place.
 
-3. Since Tom can neither easily arrange travel nor take time off from his studies in Japan, he is unable to attend his sister's wedding, thereby concluding that "Tom can't go to his sister's wedding."
-
-In summary:
-1) Tom is in Japan.
-2) Traveling to the sister's wedding location from Japan would be challenging due to various factors.
-3) Therefore, Tom cannot attend the wedding.
+3. Conclude Impossibility of Attendance: Since Tom is facing travel limitations due to his studies abroad in Japan, and the location of his sister's wedding is not in Japan, it can be concluded that he cannot attend the wedding because he physically cannot be in two places at once. Therefore, Tom will not be able to go to his sister's wedding.
